local-to-yaml.go: Implement a cache of MD5 sums
The process of calculating an MD5 checksum for all the files on what amounts to over 37 DVD-Rs of ~4.5GB each involves reading over 150GB of data.

That's rather expensive to keep repeating, so implement a persistent cache of MD5 values so that it is only necessary to calculate checksums for new files.

The cache records the MD5 sum for a file with a specific path. That's not an issue for most files as they have a fixed place on the NAS. However, .ISO files and .NRG files may appear in different locations depending on how they are mounted. That problem, if it is indeed an issue, will be handled in a future commit.