bitsavers-to-yaml.go: Ignore microfiche directory trees
Ignore everything under dec/pdp11/microfiche/Diagnostic_Program_Listings/ and dec/vax/microfiche/vms-source-listings as these are not really the manuals we are focussed on.

Restore the warning about bitsavers documents for which no MD5 is known; all documents are now listed so the warning will now be useful in the future.