file-tree-to-yaml.go: Allow YAML data to be cleared of invalid data
If a file in a directory tree is renamed or removed after a run of file-tree-to-yaml, the old file name remains behind. In the case of a rename, the "new" file has its MD5 checksum calculated but a bug meant that its filepath was left blank, leading to a fatal error.

Fix the bug but also provide a way to list YAML entries that do not have a corresponding file any more (--fnf-list) and a way to discard such obsolete entries (--fnf-discard).

This still leaves the issue of a file that is altered (e.g. it is edited or it is discarded and a file with an identical name is inserted in its place) but that can be solved by allowing for a later final (somewhat expensive, thanks to MD5) verification pass. This may be added in a later commit.