Eliminate ".." in resulting filepaths
When an index HTML file is not in the root of the target directory tree, then it is not possible to determine the absolute path to the root, and so the resulting filepath will always be relative.

To fix this allow the index file to list an optional path to the root directory.
Use this to work out the Filepath component from the root.