Fix issues found by new unit tests
SetFlags and ClearFlags need a pointer to the Document so that they can change it's Flags field.

When matching for DEC part numbers, sometimes allow "/" as a valid character. Also allow for DEC-xx... and MAINDEC-... style part numbers.

Make the date handling code interpret "00"~"25" as "2000"~"2025".

Fix the xxxFlags functions to correctly update the Document.Flags field.