Allow for DEC Printset part numbers
DEC printsets are numbered in the form "MPnnnnn", where n is a decimal digit.
Sometimes the part number is actually given as "MP-nnnnn", so allow for both.
There is sometimes an optional revision number at the end, so far this has always been in the form "-nn".

Finally, some printsets have part numbers such as "MP-0RX11-00", but this is already allowed for in the existing part number parsing.