Improve tests for indirect file parsing
The tests for valid indirect files would never have reported an error, so fix that and then fix the resulting test breakages (the code being tested was not in error, just the tests).

Add some tests for invalid indirect files.

If an error occurs, the ParseIndirectFile code needs to return an err rather than log a fatal error.