Implement a generic persistent store
The code already has a persistent store of "filepath=>MD5 checksum", which is currently called a cache. It is really a persistent store, backed by a YAML file.

This code implements a persistent store that provides the same functionality but should also provide the same functionality for the vaxhaven-to-yaml filesize store and a future "MD5 checksum => Document" store (which would record document details that can be manually updated to provide accurate dates and titles etc.)