bitsavers-to-yaml.go: Begin switch to common MD5 store
The new MD5 store now has the full set of entries for bitsavers (built from the existing bitsavers MD5 checksums).

Begin the transition to this new store by adding code which performs the required lookup against the new store and checks that the results exactly match the existing store. In the event of an issue, display a warning message.