process-digital-SOC/00-index.csv: CSV file for SOC data
The following columns exist:

PageSourceFile:
The basename of the IA page that links to a number of documents, e.g. SOHOMEHM-19991129023950.HTM. There is one such unique name for each SOC page I have found on IA.

Title:
The document title, as listed on the archived web page.

Link:
A link to an individual document: this will be either a .DOC or a .PDF (the script recognises nothing else). Each document is lsited as both a PDF and a DOC and so there should be an entry with identical details except for the link which will diff for the .PDF and the .DOC.

Date:
The date of the document, as recorded in the archived web page.

Page Count:
The page count of the linked document, as recorded in the archived web page. I've recorded this "because it's there" but I don't expect it to be useful.

In addition PageSourceFile is "abused" to allow for "Sections" and "Subsections" in the final web page.

All of this should eventually be recorded somewhere, but for now this commit message will have to serve as a reminder to me!