Move towards new INDEX CSV format
process-digital-SOC/00-index.csv: CSV file has been converted to the new format.

process-digital-SOC/build-webpage.py: Script modified to handle the new format. The checking is fairly rudimentary and there is no test code. Those deficiencies will be addressed, but for now it works in the one case where it needs to work!