Minor updates to build files
process-digital-SOC/check-csv.py: Takes a full CSV, which may include files that cannot be downloaded, and produces a CSV that only contains files that exist locally. This makes it easier to build a web page that only links to files that are available.

process-digital-SOC/build-webpage.py: Builds a web page. Includes links to the source of the files.

process-digital-SOC/convert-csv.py: A throw-away script that converted from an earlier CSV format to one that is close to the one documeted here.