Add code to produce a CSV file from YAML files
When looking for a document that may exist either locally or on the internet, rather than searching on manx and then locally, searching a single aggregated source of information would be more convenient. This program reads one or more YAML files and produces one line of CSV for each document.

yaml-to-csv/yaml-to-csv.go: Code that imports each YAML file, converts each entry to a CSV record and finally writes all the records out to a CSV file.

README.md: Brief description of the program as it stands.