vaxhaven-to-yaml/vaxhaven-to-yaml.go: Avoid embedded new lines
When transforming YAML to CSV, the number of CSV records was nearly double the number of YAML records found. Checking the CSV produced quickly found plenty of caes where a document title had an embedded new line. This, of course, broke the CSV format entirely.

The problem was with the YAMl, so this change removes a trailing new line. In passing, it also trims the title and part number.