Rename local-to-yaml to local-archive-to-yaml
Rename local-to-yaml to local-archive-to-yaml to better reflect its purpose. It is supposed to analyse documents archived on DVD-R to produce a YAML representation of them. It is not intended to analyse an otherise bare file tree that has not had some form of metadata added.

Enhance the initial comment to make it clear that it analyses an index.htm file that was produced in a certain format and will not change in the future (as the original data is on DVD-R).