local-archive-to-yaml: WIP: categorise local archive
Local archives written so far can be split into four categories and each can be properly processed just by looking at the files in the root directory.

This is a work-in-progress, being committed mostly as a backup in case of local issues.

The simplest category has an "index.htm" at the root and all documentsare directly linked from that. This category is completely processed.

The next category has an "INDEX.HTM" in the root and all links point to a further HTML file in the "HTML/" subdirectory. This category (in fact, precisely one archive) is also completely processed.

There are various problems highlighted during processing and those will be addressed in future updates.