local-archive-to-yaml.go: Add initial processing for 'custom'
The one 'custom' local archive has an index.htm that conatins both docs and further .htm files to scan (although the tree of scanning stops there ... none of these additional .htm files contain links to further .htm files that need to be scanned).

This change scans all of the .htm files that need scanning; the only things not processed are documents in the top-level index.htm file.