local-archive-to-yaml.go: Streamline indirect file
The indirect file used to list every HTML file that contained links, which for some archive volumes might mean several entries. In the event that some of these HTML files were in subdirectories of the archive volume root, third parameter provided a way to specify the root directory.

Now the first parameter is always the archive volume root and the code works out how to find all the relevant HTML files. So the third parameter is now not required.

This change removes that third parameter.

Also removed an unnecessary print statement and fixed a few typos in comments.