local-archive-to-yaml.go: One step towards correcting missing files
Some of the local archives on optical media have errors in the index HTML files. In two cases the file exists but the filename is slightly misspelled and in one case the file was unreadable and so is missing from the archive.

Rather than having yet another input file 9or files) to describe these anomalies, enhance the index file to allow for additional information such as this to be passed in.

The first step is to recognise the existing lines by processing a prefix ("archive:") and organising things so that additional prefixes can be added easily.

This commit is the first step towards doing that. More work is needed, but not today (and not tomorrow, as that's Christmas!!).