local-archive-to-yaml.go: Allow filepath transforms
Some of the local archive volumes (written to CD-R or DVD-R) have minor errors in the index files. Currently there are known to be two misspellings of filepaths. There is also one file on that media that is no longer readable and is therefore missing in the copy kept on a NAS.

To allow building a valid index of documents that are actually available locally, there needs to be a way of allowing for fixes to be made to these archives when transcribing them into a YAML representation.

This change allows an indirect-file to specify these corrections.

This is a work-in-progress, so the code currently allows for these modifications, collects them and passes them down to the code that gathers documents. However they are not actually used yet.