local-archive-to-yaml.go: Use relative path in MD5 store
local-archive-to-yaml/local-archive-to-yaml.go: Store the MD5 checksum against the volume name plus the relative filepath within that. So, for example, DEC_0026//mti/mxv22m.pdf. That should provide a unique index (volume names never repeat and filepaths withing a volume must necessarily be unique) while allowing the checksum index to be independent actual location of the volume being checked.