yaml-to-csv.go: Fix document accumulation
The statistics displayed were confusing and wrong. The first YAML file processed was always correct but the next two would produce numbers of documents read from YAML that were often larger than the known, correct value.

It turns out the code was adding documents to an existing map each time and then converting to CSV. So firstly it would read all the docs from FILE-1 and correctly handle them. Then it would read all the docs from FILE-2 and add each of those to the map that already contained docs from FILE-1. Most of these did not have clashing kyes so the total number of documents would now be too large, and (most of) the FILE-1 documents would get added to the final CSV twice.

The fix is simple enough: start with a new map for each YAML file.

Cerified that the resulting statistics are correct when specifying three YAML files as input.