INDEX-CSV.md: Add a dedicated field for MD5
No index.csv has been committed to permanent media yet, so it's fine to move the MD5 checksum from the optional field to a new mandatory field.

The only files that use the existing format are all local and all apply to pending collections fo files, so there's no problem with changing the format (as long as I upgrade the existing files).