local-archive-to-yaml.go tests
This code needs more tests. These tests also need to be run more often as the first thing I noticed is that the test of indirect file parsing is broken, so that gets commented out for now as it will take a bit longer to fix.

The test for StripOptionalLeadingAndTrailingDoubleQuotes showed that it didn't handle an empty string grafeully, so fixed that, even though it currently never needs to.

The test for TidyDocumentTitle found a few minor edge cases (again, never seen in practice) that perhaps should be adjusted, so they were adjusted.