file-tree-to-yaml.go: Write YAML data in order
Go maps are, intentionally, not ordered. I want to compare output YAML data (of a map) from run to run to check on changes. That's very awkward as even if nothing else changes, the output file does. I can get around that by using yq to sort, but that's hassle at best.

So marshall the data one entry at time in the required order. This may not be the most efficient solution, but it works and for the amount of data I have, it's perfectly functional.