local-archive-check.go: Only check essential files that exist
There's no point reporting that, for example, every file in the tree is not present in the md5sum file if the md5sum itself does not exist. So avoid those checks for essential metafiles that do not exist: their absence will trigger a fatal error anyway.

Tidy up FATAL lines and make sure that non-fatal ones say INFO just to make the output easier to read at a glance.